Consolidate Infested Blocks into a single file
